Read in UM_IC

+++ include/TestProb.h
@@ -47,7 +47,7 @@ extern void (*Par_Init_Attribute_User_Ptr)();
 #ifdef GRAVITY
 #include "CUPOT.h"
 extern double ExtPot_AuxArray[EXT_POT_NAUX_MAX];
-extern double ExtAcc_AuxArray[EXT_ACC_NAUX_MAX];
+extern float ExtAcc_AuxArray[EXT_ACC_NAUX_MAX];
 #endif
